move passthrough  and dummy structs to signer_core as test-utils feature
