+++ confidential_signer_mock/Cargo.toml
@@ -4,12 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-# Workspace/local dependencies
 confidential_signer = { path = "../confidential_signer" }
-
-# CLI
+signer_core = { path = "../signer_core", features = ["test-utils"] }
 clap = { version = "4.5", features = ["derive"] }
 clap_derive = "4.5"
-
-# Serialization
-serde = { version = "1.0", features = ["derive"] }
