+++ confidential_signer_mock/src/app.rs
@@ -1,10 +1,6 @@
-use confidential_signer::{
-    rand_core,
-    signer_core::{EncryptionBackend, EncryptionBackendFactory, rpc::server::Server},
-    tokio,
-};
-use serde::{Deserialize, Serialize};
-use std::{convert::Infallible, io, net::SocketAddr};
+use confidential_signer::{rand_core, tokio};
+use signer_core::{mock::PassthroughFactory, rpc::server::Server};
+use std::{io, net::SocketAddr};
 
 pub struct App {}
 
@@ -27,46 +23,14 @@ impl std::fmt::Display for Error {
     }
 }
 
-#[derive(Debug)]
-struct Passthrough;
-
-impl EncryptionBackend for Passthrough {
-    type Error = Infallible;
-
-    async fn encrypt(&self, src: &[u8]) -> Result<Vec<u8>, Self::Error> {
-        Ok(Vec::from(src))
-    }
-
-    async fn decrypt(&self, src: &[u8]) -> Result<Vec<u8>, Self::Error> {
-        Ok(Vec::from(src))
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-struct DummyCredentials {} // serialized as empty object instead of null for unity
-
-struct PassthroughFactory;
-
-impl EncryptionBackendFactory for PassthroughFactory {
-    type Output = Passthrough;
-    type Credentials = DummyCredentials;
-    async fn try_new(&self, _cred: Self::Credentials) -> Result<Self::Output, Infallible> {
-        Ok(Passthrough)
-    }
-}
-
 impl App {
     pub async fn run(addr: &SocketAddr) -> Result<(), Error> {
-        // Start listening
         let listener = tokio::net::TcpListener::bind(addr).await?;
         loop {
-            // Accept TCP connection
             let (conn, _) = listener.accept().await?;
-
             tokio::spawn(async move {
-                // Create server
                 let mut srv = Server::new(PassthroughFactory, rand_core::OsRng);
-                // Serve connection
+
                 if let Err(err) = srv.serve_connection(conn).await {
                     eprintln!("{}", err);
                 }
